Add access type to series reservation in admin panel
